+++ src/app/components/NavList.tsx
@@ -40,8 +40,10 @@ export default function NavList({
                 ) : (
                   <>{item.name}</>
                 )}
-                {mainMenu ?? (
+                {mainMenu ? (
                   <span className="absolute -bottom-2 left-1/2 h-0.5 w-0 -translate-x-1/2 bg-black transition-all duration-300 group-hover:w-full"></span>
+                ) : (
+                  <></>
                 )}
               </Link>
             ) : (
